analysis/land_segmentation/requirements.txt
albucore==0.0.17
albumentations==0.5.2
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.3.0
babel==2.17.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bash-kernel==0.10.0
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach==6.1.0
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
charset-normalizer==3.4.4
comm==0.2.3
conda==4.9.2
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
contourpy==1.1.1
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.12.1
debugpy==1.8.17
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml==0.7.1
dnspython==2.1.0
efficientnet-pytorch==0.6.3
eval-type-backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
filetype==1.2.0
fonttools==4.57.0
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.35.1
imgaug==0.4.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==8.1.7
jedi==0.17.0
jinja2==3.1.6
json5==0.12.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.10.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.3.0
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.8
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kagglehub==0.2.9
kiwisolver==1.4.7
lazy-loader==0.4
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.1.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
munch==4.0.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.4.7
notebook-shim==0.2.4
numpy==1.20.3
olefile==0.46
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
overrides==7.7.0
packaging==25.0
pandas==2.0.0
pandocfilters==1.5.1
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
pkginfo==1.7.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
pretrainedmodels==0.7.4
prometheus-client==0.21.1
prompt-toolkit==3.0.52
psutil @ file:///tmp/build/80754af9/psutil_1612298023621/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.3
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==2.10.6
pydantic-core==2.27.2
Pygments @ file:///tmp/build/80754af9/pygments_1613849567472/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
python-etcd==0.4.5
python-json-logger==4.0.0
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==27.1.0
referencing==0.35.1
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
ruamel-yaml==0.15.87
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
segmentation-models-pytorch==0.1.2
Send2Trash==1.8.3
shapely==2.0.7
six @ file:///tmp/build/80754af9/six_1605205327372/work
sniffio==1.3.1
soupsieve @ file:///tmp/build/80754af9/soupsieve_1614023247537/work
stack-data==0.6.3
terminado==0.18.1
tifffile==2023.7.10
timm==0.1.20
tinycss2==1.2.1
tomli==2.3.0
torch==1.8.1
torchelastic==0.2.2
torchtext==0.9.1
torchvision==0.9.1
tornado==6.4.2
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets==5.14.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
zipp==3.20.2
